refactor: simulator state/render in separate files
closes #12
