+++ include/simulator/simulation_config.h
@@ -1,0 +1,11 @@
+#ifndef SIMULATION_CONFIG
+#define SIMULATION_CONFIG
+
+// TODO: Make this configurable
+constexpr int GRID_CENTER = 500;
+constexpr int GRID_SIZE = 1000;
+constexpr int SQUARE_PIXEL_DIM = 100;
+constexpr int SQUARE_RESIZE = 64;
+constexpr int SCREEN_CENTER = GRID_CENTER;
+
+#endif
